refactor: 일급 컬렉션 변경 및 client -> provider로 통일
